# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CaptchaVerify:
    properties:
      code:
        type: string
      id:
        type: string
    required:
    - code
    - id
    type: object
...
  echox.Response:
    properties:
      data:
        type: object
      message:
        type: object
    type: object
...
info:
  contact:
    email: liusha@email.cn
    name: LiuSha
paths:
...
  /api/publics/captcha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Get Captcha
      tags:
      - Public
  /api/publics/captcha/verify:
    post:
      parameters:
      - description: CaptchaVerify
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CaptchaVerify'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: Verify Captcha
      tags:
      - Public
...
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"